Update Swin Transformer for mmdet's fp16

+++ mmdet/models/backbones/swin_transformer.py
@@ -179,7 +179,7 @@ class WindowAttention(nn.Module):
 
         attn = self.attn_drop(attn)
 
-        x = (attn @ v).transpose(1, 2).reshape(B_, N, C)
+        x = (attn.type_as(v) @ v).transpose(1, 2).reshape(B_, N, C)
         x = self.proj(x)
         x = self.proj_drop(x)
         return x
